+++ packages/components/CHANGELOG.md
@@ -9,6 +9,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### [Unreleased]
 
+#### Added
+
+- `xxl` breakpoint for all breakpoint customizable component values [PR](https://github.com/porsche-design-system/porsche-design-system/pull/2454)
+
 #### Fixed
 
 - Disabled color of `Icon` component [PR](https://github.com/porsche-design-system/porsche-design-system/pull/2446)